pyproject.toml
[tool.poetry]
name = "matterless"
version = "0.0.0"
description = "Matterless"
authors = ["Mark Gomersbach <markgomersbach@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mgomersbach/matterless"
repository = "https://github.com/mgomersbach/matterless"
documentation = "https://matterless.readthedocs.io"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.urls]
Changelog = "https://github.com/mgomersbach/matterless/releases"

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.0.1"
mattermostdriver = "^7.3.2"
appdirs = "^1.4.4"
shelved-cache = "^0.2.1"
pyTermTk = "^0.9.0-alpha.62"
pylint = "^2.14.4"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
nox-poetry = "^1.0.0"

[tool.poetry.scripts]
matterless = "matterless.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["matterless", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
preview = true
